modify models id int to int64

+++ pkg/models/audit.go
@@ -4,9 +4,9 @@ import "yuudidi.com/pkg/utils"
 
 type AuditMessage struct {
 	ID
-	MerchantId    int `gorm:"index:IDX_MERCHANT" json:"merchant_id"`
-	DistributorId int `gorm:"index:IDX_DISTRIBUTOR" json:"distributor_id"`
-	OperatorId    int `gorm:"type:int(11)" json:"operator_id"`
+	MerchantId    int64 `gorm:"index:IDX_MERCHANT" json:"merchant_id"`
+	DistributorId int64 `gorm:"index:IDX_DISTRIBUTOR" json:"distributor_id"`
+	OperatorId    int64 `gorm:"type:int(11)" json:"operator_id"`
 	//联系电话
 	ContactPhone string `gorm:"type:varchar(20)" json:"contact_phone"`
 	//原因
